requirements.txt
datasets==2.3.2
faiss-cpu==1.7.2 #faiss-gpu==1.7.2  # to compute fid features
huggingface-hub==0.12.0     # to have access to comet22-kiwi
jieba==0.42.1   # to compute fid features
matplotlib==3.5.1
multiprocess==0.70.13
nltk==3.7       # to compute fid features
numpy==1.22.1
pandas==1.5.3
PyYAML==6.0.2
rapidfuzz==3.0.0    # to compute fid features
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentencepiece==0.1.96
sentence-transformers==2.2.2    # to compute fid features
statsmodels==0.13.2
torch==2.1.1
torchvision==0.16.1
transformers==4.19.4
unbabel-comet==2.0.1 
xgboost==1.5.2
